vboot {BootValidation}	R Documentation

Generic function for bootstrap validation

Description

Validate 'glmnet' linear, logistic or cox regression using bootstrap.

Usage

vboot(fit, x, y, s, nfolds = 5, B = 200, cv_replicates = 100,
  n_cores = max(1, parallel::detectCores() - 1))

Arguments

fit
Object from glmnet fit

x
A matrix of the predictors, each row is an observation vector.

y
A vector of response variable. It should be quantitative for lineal regression, a factor with two levels for logistic regression or a two-column matrix with columns named 'time' and 'status' for cox regression.

s
Value of the penalty parameter "lambda" selected from the original 'cv.glmnet'.

nfolds
Number of folds for cross validation as in 'cv.glmnet'.

B
Number of bootsrap samples

cv_replicates
Number of replicates for the cross-validation step

n_cores
number of cores to use in parallel. Default detectCores()-1

References

Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software, 33(1), 1-22. URL http://www.jstatsoft.org/v33/i01/.

Noah Simon, Jerome Friedman, Trevor Hastie, Rob Tibshirani (2011). Regularization Paths for Cox's Proportional Hazards Model via Coordinate Descent. Journal of Statistical Software, 39(5), 1-13. URL http://www.jstatsoft.org/v39/i05/.

Harrell Jr, F. E. (2015). Regression modeling strategies: with applications to linear models, logistic and ordinal regression, and survival analysis. Springer.

Gordon C.S. Smith, Shaun R. Seaman, Angela M. Wood, Patrick Royston, Ian R. White (2014). Correcting for Optimistic Prediction in Small Data Sets, American Journal of Epidemiology, Volume 180, Issue 3, 1 August 2014, Pages 318-324, https://doi.org/10.1093/aje/kwu140

Examples

# Create the data
set.seed(25)
x <- matrix(rnorm(80),ncol=4)
y <- x[,4]*0.8+x[,3]*0.4+rnorm(20)
# Fit glmnet model
fit_enet <- glmnet::glmnet(x, y, alpha = 0.5)
# Bootstrap validation
vboot(fit_enet, x, y, nfolds = 3, B = 2, s = 0.5, cv_replicates = 5, n_cores = 1)

[Package BootValidation version 0.1.65 Index]