vboot.multnet {BootValidation}	R Documentation

Internal bootstraping validation multinomial glmnet model

Description

Validate glmnet logistic regression using bootstrap.

Usage

## S3 method for class 'multnet'
vboot(fit, x, y, s, nfolds = 5, B = 200,
  cv_replicates = 100, n_cores = max(1, parallel::detectCores() - 1))

Arguments

fit
Object from glmnet fit

x
A matrix of the predictors, each row is an observation vector.

y
A vector of response variable. Should be a factor with two levels

s
Value of the penalty parameter "lambda" selected from the original 'cv.glmnet'

nfolds
Number of folds for cross validation as in cv.glmnet

B
Number of bootsrap samples

cv_replicates
Number of replicates for the cross-validation step in 'cv.glmnet'

n_cores
number of cores to use in parallel. Default detectCores()-1

[Package BootValidation version 0.1.65 Index]